+++ src/elias.h
@@ -1,9 +1,10 @@
 #include <stdlib.h>
+#include <stdint.h>
 #include "bitlevel_io.h"
 #include <math.h>
 #ifndef ELIAS_CODE
 #define ELIAS_CODE
-struct bitlevel_object calculate_elias_value(int value);
-void write_elias_value(struct bitlevel_file_pointer * file_pointer, int value);
-int read_elias_value(struct bitlevel_file_pointer * file_pointer);
+struct bitlevel_object calculate_elias_value(uint64_t value);
+void write_elias_value(struct bitlevel_file_pointer * file_pointer, uint64_t value);
+uint64_t read_elias_value(struct bitlevel_file_pointer * file_pointer);
 #endif
